Suggest Backbone
So after looking at Spine, Soma.js, Batman, CanJS, and Angular. I think that using Backbone may be a good idea... Heres why
Spine: Prefers coffeescript, and all of the documentation is written as such. This is a bit of a deal breaker considering we are trying to teach javascript and coffeescript is another beast entirely.
Soma.Js: I really liked this one and it was *tiny* really really tiny, but it seemed to be a bit more useful if you had some background. It gave you primitives but not much in the way of suggestions. 
Batman: Gotta love coffeescript
CanJS: You have to roll your own two way bindings, and it is not well known in industry
Backbone: Well known by many people, the primitives it provides are in all the frameworks, and it has annotated source code available as well as annotated example apps.